\name{bering.repo-package}
\alias{bering.repo-package}
\alias{bering.repo}
\docType{package}
\title{
\packageTitle{bering.repo}
}
\description{
\packageDescription{bering.repo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bering.repo}
\packageIndices{bering.repo}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bering.repo}

Maintainer: \packageMaintainer{bering.repo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
